+++ java/opendcs/src/main/java/decodes/dbeditor/ConfigSelectPanel.java
@@ -1,50 +1,38 @@
 /*
-*  $Id$
-*
-*  $Log$
-*  Revision 1.9  2011/01/07 16:01:56  mmaloney
-*  bugfixes
-*
-*  Revision 1.8  2010/09/16 15:03:13  mmaloney
-*  dev
-*
-*  Revision 1.7  2010/09/16 14:49:11  mmaloney
-*  double-click code causing exception. Removed for now.
-*
-*  Revision 1.6  2009/08/12 19:54:57  mjmaloney
-*  usgs merge
-*
-*  Revision 1.5  2008/12/14 00:51:28  mjmaloney
-*  platform count fix
-*
-*  Revision 1.4  2008/11/20 18:49:21  mjmaloney
-*  merge from usgs mods
-*
-*/
+ * Where Applicable, Copyright 2025 OpenDCS Consortium and/or its contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy
+ * of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
 package decodes.dbeditor;
 
 import java.awt.*;
 import java.awt.event.*;
-import java.util.Vector;
-import java.util.Iterator;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.Comparator;
-import java.util.Enumeration;
 import java.util.ResourceBundle;
 import javax.swing.*;
 
 import ilex.util.TextUtil;
 import decodes.db.*;
 import decodes.gui.*;
+import org.slf4j.LoggerFactory;
 
 /**
 Panel for selecting a Platform Configuration. Used inside ConfigSelectDialog.
 */
 public class ConfigSelectPanel extends JPanel
 {
-	static ResourceBundle genericLabels = DbEditorFrame.getGenericLabels();
 	static ResourceBundle dbeditLabels = DbEditorFrame.getDbeditLabels();
+	private static final org.slf4j.Logger log = LoggerFactory.getLogger(ConfigSelectPanel.class);
 
 	BorderLayout borderLayout1 = new BorderLayout();
 	JScrollPane jScrollPane1 = new JScrollPane();
@@ -81,7 +69,7 @@ public class ConfigSelectPanel extends JPanel
 		    jbInit();
 		}
 		catch(Exception ex) {
-		    ex.printStackTrace();
+		    log.atInfo().log("Error initializing ConfigSelectPanel",ex);
 		}
 	}
 
@@ -186,114 +174,6 @@ public class ConfigSelectPanel extends JPanel
 	}
 }
 
-class ConfigSelectTableModel extends javax.swing.table.AbstractTableModel
-	implements SortingListTableModel
-{
-	private String colNames[] = 
-	{ 
-		ConfigSelectPanel.dbeditLabels.getString("ConfigSelectPanel.col0"),
-		ConfigSelectPanel.dbeditLabels.getString("ConfigSelectPanel.col1"),
-		ConfigSelectPanel.dbeditLabels.getString("ConfigSelectPanel.col2"),
-		ConfigSelectPanel.dbeditLabels.getString("ConfigSelectPanel.col3")
-	};
-	private ConfigSelectPanel panel;
-	private Vector vec;
-	private int lastSortColumn;
-
-	public ConfigSelectTableModel(ConfigSelectPanel ssp)
-	{
-		super();
-		lastSortColumn = -1;
-		this.panel = ssp;
-		refill();
-		this.sortByColumn(0);
-	}
-
-	void refill()
-	{
-		try 
-		{
-			Database.getDb().platformConfigList.read();
-			Database.getDb().platformConfigList.countPlatformsUsing();
-		} catch (DatabaseException dbe ) { };
-		vec = new Vector(Database.getDb().platformConfigList.values());
-	}
-
-	void add(PlatformConfig ob)
-	{
-		for(int i=0; i<vec.size(); i++)
-		{
-			PlatformConfig pc = (PlatformConfig)vec.elementAt(i);
-			if (pc.configName.equals(ob.configName))
-				return;
-		}
-		vec.add(ob);
-		fireTableDataChanged();
-	}
-
-	void replace(DatabaseObject oldOb, DatabaseObject newOb)
-	{
-		vec.remove(oldOb);
-		vec.add(newOb);
-		if (lastSortColumn != -1)
-			sortByColumn(lastSortColumn);
-		else
-			fireTableDataChanged();
-	}
-
-	void deleteAt(int index)
-	{
-		PlatformConfig ob = (PlatformConfig)vec.elementAt(index);
-		vec.remove(index);
-		try { Database.getDb().getDbIo().deleteConfig(ob); }
-		catch(DatabaseException e)
-		{
-			TopFrame.instance().showError(e.toString());
-		}
-		Database.getDb().platformConfigList.remove(ob);
-		fireTableDataChanged();
-	}
-
-	public int getColumnCount()
-	{
-		int r = colNames.length;
-		return r;
-	}
-
-	public String getColumnName(int c)
-	{
-		return colNames[c];
-	}
-
-	public boolean isCellEditable(int r, int c) { return false; }
-
-	public int getRowCount()
-	{
-		return vec.size();
-	}
-
-	public Object getValueAt(int r, int c)
-	{
-		return ConfigColumnizer.getColumn(getConfigAt(r), c);
-	}
-
-	PlatformConfig getConfigAt(int r)
-	{
-		return (PlatformConfig)getRowObject(r);
-	}
-
-	public Object getRowObject(int r)
-	{
-		return vec.elementAt(r);
-	}
-
-	public void sortByColumn(int c)
-	{
-		lastSortColumn = c;
-		Collections.sort(vec, new pcColumnComparator(c));
-		fireTableDataChanged();
-	}
-}
 
 class ConfigColumnizer
 {
